chore(app): tagging version v1.8.7

+++ main.go
@@ -3,15 +3,13 @@ package main
 import (
 	"embed"
 	"fmt"
-	"github.com/razshare/frizzante/fs"
+	ffs "github.com/razshare/frizzante/fs"
 	flag "github.com/spf13/pflag"
+	"io"
+	"io/fs"
 	"log"
 	"os"
-	"path/filepath"
 )
-
-const binaryVersion = "1.8.6"
-const projectVersion = "1.2.8"
 
 var FlagHelp = flag.BoolP("help", "h", false, "shows the help document")
 var FlagVersion = flag.BoolP("version", "v", false, "shows the binary version and the project version")
@@ -19,15 +17,26 @@ var FlagCreateProject = flag.StringP("create-project", "c", "", fmt.Sprintf("cre
 var FlagRestoreUtilities = flag.StringP("restore-utilities", "r", "", fmt.Sprintf("restores frizzante utilities"))
 
 //go:embed app/lib/utilities
+//go:embed project.zip
+//go:embed version
 var embedded embed.FS
 
 func main() {
 	flag.Parse()
 
+	var version string
+
+	versionData, versionError := embedded.ReadFile("version")
+	if versionError != nil {
+		log.Fatal(versionError)
+	}
+
+	version = string(versionData)
+
 	if *FlagRestoreUtilities != "" {
 		var restore func(from string, to string)
 		restore = func(from string, to string) {
-			if !fs.IsDirectory(to) {
+			if !ffs.IsDirectory(to) {
 				mkdirAllError := os.MkdirAll(to, os.ModePerm)
 				if mkdirAllError != nil {
 					log.Fatal(mkdirAllError)
@@ -67,7 +76,7 @@ func main() {
 	}
 
 	if *FlagVersion {
-		fmt.Printf("v%s using project v%s\n", binaryVersion, projectVersion)
+		println(version)
 		os.Exit(0)
 	}
 
@@ -77,27 +86,39 @@ func main() {
 	}
 
 	if *FlagCreateProject != "" {
-		downloadError := fs.DownloadFile(fmt.Sprintf("https://github.com/razshare/frizzante-starter/archive/refs/tags/v%s.zip", projectVersion), *FlagCreateProject+".zip")
-		if downloadError != nil {
-			log.Fatal(downloadError)
+		srcFile, srcError := embedded.Open("project.zip")
+		if srcError != nil {
+			log.Fatal(srcError)
+		}
+		defer func(srcFile fs.File) {
+			err := srcFile.Close()
+			if err != nil {
+				log.Fatal(err)
+			}
+		}(srcFile)
+
+		destFile, destError := os.Create(*FlagCreateProject + ".zip")
+		if destError != nil {
+			log.Fatal(destError)
+		}
+		defer func(destFile *os.File) {
+			err := destFile.Close()
+			if err != nil {
+				log.Fatal(err)
+			}
+		}(destFile)
+
+		_, copyError := io.Copy(destFile, srcFile)
+		if copyError != nil {
+			return
 		}
 
-		unzipError := fs.UnzipFile(*FlagCreateProject+".zip", "."+*FlagCreateProject+".tmp")
+		unzipError := ffs.UnzipFile(*FlagCreateProject+".zip", *FlagCreateProject)
 		if unzipError != nil {
 			log.Fatal(unzipError)
 		}
 
 		removeError := os.Remove(*FlagCreateProject + ".zip")
-		if removeError != nil {
-			log.Fatal(removeError)
-		}
-
-		renameError := os.Rename(filepath.Join("."+*FlagCreateProject+".tmp", fmt.Sprintf("frizzante-starter-%s", projectVersion)), *FlagCreateProject)
-		if renameError != nil {
-			log.Fatal(renameError)
-		}
-
-		removeError = os.RemoveAll("." + *FlagCreateProject + ".tmp")
 		if removeError != nil {
 			log.Fatal(removeError)
 		}
